Move translations into api proper
