--- text-feature-translation/build.gradle
@@ -1,9 +1,0 @@
-dependencies {
-  api project(':adventure-api')
-}
-
-jar {
-  manifest.attributes(
-    'Automatic-Module-Name': 'net.kyori.adventure.text.feature.translation'
-  )
-}
